+++ frontend/appflowy_flutter/lib/plugins/database/application/cell/select_option_cell_service.dart
@@ -1,5 +1,5 @@
 import 'package:appflowy/plugins/database/application/field/type_option/type_option_service.dart';
-import 'package:appflowy_backend/protobuf/flowy-database2/select_option.pb.dart';
+import 'package:appflowy_backend/protobuf/flowy-database2/select_option_entities.pb.dart';
 import 'package:dartz/dartz.dart';
 import 'package:appflowy_backend/dispatch/dispatch.dart';
 import 'package:appflowy_backend/protobuf/flowy-error/errors.pb.dart';
