+++ frontend/appflowy_flutter/lib/plugins/database/application/field/type_option/single_select_type_option.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:appflowy/plugins/database/grid/presentation/widgets/header/type_option/builder.dart';
 import 'package:appflowy_backend/log.dart';
-import 'package:appflowy_backend/protobuf/flowy-database2/select_option.pb.dart';
+import 'package:appflowy_backend/protobuf/flowy-database2/select_option_entities.pb.dart';
 
 import 'select_option_type_option_bloc.dart';
 import 'type_option_service.dart';
