+++ frontend/rust-lib/flowy-database2/src/entities/type_option_entities/mod.rs
@@ -2,7 +2,7 @@ mod checkbox_entities;
 mod checklist_entities;
 mod date_entities;
 mod number_entities;
-mod select_option;
+mod select_option_entities;
 mod text_entities;
 mod timestamp_entities;
 mod url_entities;
@@ -11,7 +11,7 @@ pub use checkbox_entities::*;
 pub use checklist_entities::*;
 pub use date_entities::*;
 pub use number_entities::*;
-pub use select_option::*;
+pub use select_option_entities::*;
 pub use text_entities::*;
 pub use timestamp_entities::*;
 pub use url_entities::*;
