+++ frontend/rust-lib/flowy-database2/src/services/field/type_options/checklist_type_option/mod.rs
@@ -1,5 +1,6 @@
 mod checklist;
 mod checklist_entities;
+mod checklist_filter;
 
 pub use checklist::*;
 pub use checklist_entities::*;
