+++ frontend/rust-lib/flowy-database2/src/services/field/type_options/selection_type_option/mod.rs
@@ -1,4 +1,3 @@
-mod checklist_filter;
 mod multi_select_type_option;
 mod select_filter;
 mod select_ids;
